fix benchmark for working with tests which pass less then 1 ms.

+++ src/main/java/com/antkorwin/junit5integrationtestutils/test/extensions/benchmark/TestTiming.java
@@ -19,7 +19,7 @@ import java.util.OptionalDouble;
 public class TestTiming {
 
     private long startTime;
-    private long duration;
+    private double duration;
     private double average;
     private List<Long> results = new ArrayList<>();
